refactor: restructure site and root layouts
