+++ src/app/(site)/chronik/[showId]/page.tsx
