+++ src/app/(site)/chronik/fullframes.tsx
