+++ src/app/(site)/chronik/poster-slideshow.tsx
