+++ src/app/(site)/mystery/_components/mystery-countdown-card.tsx
