+++ src/app/(site)/mystery/_components/mystery-scoreboard.tsx
