+++ src/app/(site)/mystery/_components/mystery-tips-board.tsx
