+++ src/app/(site)/mystery/page.tsx
