+++ src/app/(site)/unsere-schulkatze/image-rotator.tsx
